fix: code hygiene — bugs, leaks, boxing, and LINQ inefficiencies
Bug fixes:
- Fix callback typo EnablePropertyMangagerPage → nameof()
- Fix GetCutListFolder guard (|| → &&) always returning null
- Fix SubmitSelectionCallback always rejecting (bool res = false → true)
- Fix PreviousPageClicked routing to wrong event source
- Fix Mathematics.AlmostEqual ignoring digits parameter
- Fix OnStart events firing twice then clearing

Resource leak fixes:
- Wrap Graphics/ImageAttributes in using (MaskedBitmap)
- Dispose maskImage/maskedImage on exception paths
- Add Marshal.ReleaseComObject in feature/sub-feature traversal loops
- Wrap RegistryKey in using statements
- Dispose ResourceSet in AddinIcons.GetResxBitmap
- PopupWinForm now properly implements IDisposable
- Null out PmpBitmap._bitmap after disposal
- Release COM intermediaries in Face2Extensions

Performance fixes:
- Replace HasFlag() with bitwise check (avoids boxing)
- Replace bool.ToString().GetHashCode() with direct int hashing
- Replace .Cast<int>() with .Select(f => (int)f) on enums
- Replace .Count() with .Count on List/IList (6 sites)
- Replace .ElementAt(i) with indexer after .ToList() (8 sites)
- Cast enum to int before GetHashCode()

Cleanup:
- Rename WeakEventSource → EventSource (was never weak)
- Remove dead Hashtable field, NotImplementedException methods
- Remove redundant try-catch-throw (5 sites)
- Remove unused usings (EnvDTE, System.Net, etc.)
- Deduplicate Units.ConvertLength string overload
- Replace throw e with throw (4 sites)
- Add Debug.WriteLine to empty catch blocks in RegistryHelper

+++ Addins/UI/ContextMenus/ContextMenuRegistrar.cs
@@ -376,7 +376,7 @@ namespace Hymma.Solidworks.Addins.ContextMenus
                 unchecked
                 {
                     var hash = 17;
-                    hash = (hash * 31) + _selectionType.GetHashCode();
+                    hash = (hash * 31) + ((int)_selectionType).GetHashCode();
                     hash = (hash * 31) + StringComparer.OrdinalIgnoreCase.GetHashCode(_title);
                     return hash;
                 }
